Created top level JS files for realms. Fixed checkRedirect() bug
