+++ src/js/server/admin/admin-realm.js
@@ -9,11 +9,17 @@ const AdminRealm = {
   checkRedirect() {
     if (!ServerIp.isSet()) {
       m.route('/');
+
+      return true;
     }
 
     if (!JwtAuth.isAuthenticated()) {
       m.route('/server/admin/login');
+
+      return true;
     }
+
+    return false;
   },
   constructView(childContent) {
     return [AdminMenu, m('div', childContent)];
