+++ src/js/site/site.js
@@ -1,0 +1,13 @@
+import Home from './home';
+import About from './about';
+
+const Site = {
+  getRoutes() {
+    return {
+      '/': Home,
+      '/about': About
+    };
+  }
+};
+
+export default Site;
